third_party/WebKit/LayoutTests/fast/canvas/canvas-lineDash-input-sequence-expected.txt

Test that setLineDash converts input argument into a Web IDL sequence

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* Test passing a Array as input.
PASS lineDash[0] is 5
PASS lineDash[1] is 15
PASS lineDash[2] is 25
* Test passing a Int8Array as input.
PASS lineDash[0] is 5
PASS lineDash[1] is 15
PASS lineDash[2] is 25
* Test passing a Int16Array as input.
PASS lineDash[0] is 5
PASS lineDash[1] is 15
PASS lineDash[2] is 25
* Test passing a Int32Array as input.
PASS lineDash[0] is 5
PASS lineDash[1] is 15
PASS lineDash[2] is 25
* Test passing a Uint8Array as input.
PASS lineDash[0] is 5
PASS lineDash[1] is 15
PASS lineDash[2] is 25
* Test passing a Uint16Array as input.
PASS lineDash[0] is 5
PASS lineDash[1] is 15
PASS lineDash[2] is 25
* Test passing a Uint32Array as input.
PASS lineDash[0] is 5
PASS lineDash[1] is 15
PASS lineDash[2] is 25
* Test passing a Float32Array as input.
PASS lineDash[0] is 5
PASS lineDash[1] is 15
PASS lineDash[2] is 25
* Test passing a Float64Array as input.
PASS lineDash[0] is 5
PASS lineDash[1] is 15
PASS lineDash[2] is 25
* Test passing a Uint8ClampedArray as input.
PASS lineDash[0] is 5
PASS lineDash[1] is 15
PASS lineDash[2] is 25
* Test passing a Object as input.
PASS lineDash[0] is 5
PASS lineDash[1] is 15
PASS lineDash[2] is 25
* Test passing a Date as input.
PASS ctx.setLineDash(inputArray) threw exception TypeError: Failed to execute 'setLineDash' on 'CanvasRenderingContext2D': The 1st argument is neither an array, nor does it have indexed properties..
* Test passing a RegExp as input.
PASS ctx.setLineDash(inputArray) threw exception TypeError: Failed to execute 'setLineDash' on 'CanvasRenderingContext2D': The 1st argument is neither an array, nor does it have indexed properties..
* Test passing an Object without length as input.
PASS ctx.setLineDash(inputArray) threw exception TypeError: Failed to execute 'setLineDash' on 'CanvasRenderingContext2D': The 1st argument is neither an array, nor does it have indexed properties..
* Test passing a Number as input.
PASS ctx.setLineDash(inputArray) threw exception TypeError: Failed to execute 'setLineDash' on 'CanvasRenderingContext2D': The 1st argument is neither an array, nor does it have indexed properties..
* Test passing a String as input.
PASS ctx.setLineDash(inputArray) threw exception TypeError: Failed to execute 'setLineDash' on 'CanvasRenderingContext2D': The 1st argument is neither an array, nor does it have indexed properties..
* Test passing a Boolean as input.
PASS ctx.setLineDash(inputArray) threw exception TypeError: Failed to execute 'setLineDash' on 'CanvasRenderingContext2D': The 1st argument is neither an array, nor does it have indexed properties..
* Test passing null as input.
PASS ctx.setLineDash(inputArray) threw exception TypeError: Failed to execute 'setLineDash' on 'CanvasRenderingContext2D': The 1st argument is neither an array, nor does it have indexed properties..
* Test passing undefined as input.
PASS ctx.setLineDash(inputArray) threw exception TypeError: Failed to execute 'setLineDash' on 'CanvasRenderingContext2D': The 1st argument is neither an array, nor does it have indexed properties..
PASS successfullyParsed is true

TEST COMPLETE
